feat: Spinner 정렬 기능 추가
