--- frontend/src/components/CenteredSpinner/CenteredSpinner.styles.tsx
@@ -1,11 +1,0 @@
-import styled from '@emotion/styled';
-
-const StyledContainer = styled.div`
-  width: 100%;
-  height: 100%;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-`;
-
-export { StyledContainer };
